Move ElementName[Component] out of core package
