+++ subprojects/core-model/src/main/java/dev/nokee/model/internal/ElementName.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.model.internal.core;
+package dev.nokee.model.internal;
 
 import lombok.EqualsAndHashCode;
