+++ subprojects/core-model/src/main/java/dev/nokee/model/internal/ElementNameComponent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.model.internal.core;
+package dev.nokee.model.internal;
+
+import dev.nokee.model.internal.core.ModelComponent;
 
 public final class ElementNameComponent implements ModelComponent {
 	private final ElementName value;
