+++ subprojects/core-model/src/main/java/dev/nokee/model/internal/registry/DefaultModelRegistry.java
@@ -23,7 +23,7 @@ import dev.nokee.model.internal.core.BindManagedProjectionService;
 import dev.nokee.model.internal.core.Bits;
 import dev.nokee.model.internal.core.DescendantNodes;
 import dev.nokee.model.internal.core.DisplayNameComponent;
-import dev.nokee.model.internal.core.ElementNameComponent;
+import dev.nokee.model.internal.ElementNameComponent;
 import dev.nokee.model.internal.core.HasInputs;
 import dev.nokee.model.internal.core.ModelAction;
 import dev.nokee.model.internal.core.ModelActionWithInputs;
